feat(state): :sparkles: move all logic from questionCard to appState
