+++ src/lib/utils/getContent.svelte.ts
@@ -1,13 +1,6 @@
 import { getLatestResponses } from '$lib/services/database/responses';
 import { getLatestActions } from '$lib/services/database';
-
-export interface QuestionDetails {
-	responseInput: string | null;
-	actionsInput: string | null;
-	actionType: string;
-	responseId: string | null;
-}
-
+import type { QuestionDetails } from '$lib/types/appState';
 
 export const getQuestionDetails = async (
 	user_id: string,
@@ -31,5 +24,6 @@ const getActionDetails = async (response_Id: string | undefined, user_id: string
 		const actionResponse = response.data.find((r) => r.response_id === response_Id);
 		return actionResponse;
 	}
+	
 	return null;
 };
